[stable @ bec7e362db719b195e82725791f27c53f5e54030] A patch by Nuno Lucas, ml-colinux at xpto.ath.cx:
When I made the patch for not letting the colinux-daemon.exe abort on
Ctr-C or the user clicking the close button (leaving the fs corrupted),
I thought the signal for user logout was not received by the daemon if
running as service.

This fixes that, so the service doesn't shutdown on user logout now.


git-svn-id: svn://svn.code.sf.net/p/colinux/code/branches/stable@195 688f6386-cf32-0410-9ead-a596f0484e40